Organize network constants, add ResponseCode enum
