--- src/network/libp2p/constants.ts
@@ -1,1 +1,0 @@
-export const RPC_MULTICODEC = "/eth/serenity/beacon/rpc/1";
