+++ src/network/libp2p/network.ts
@@ -10,9 +10,10 @@ import Gossipsub from "libp2p-gossipsub";
 import PeerInfo from "peer-info";
 
 import {Attestation, BeaconBlock, Shard, RequestBody, ResponseBody} from "../../types";
-import {BLOCK_TOPIC, ATTESTATION_TOPIC, SHARD_SUBNET_COUNT} from "../../constants";
+import {
+  Method, RequestId, BLOCK_TOPIC, ATTESTATION_TOPIC, SHARD_SUBNET_COUNT,
+} from "../../constants";
 import {INetwork, INetworkOptions} from "../interface";
-import {Method, RequestId} from "../codec";
 import {shardAttestationTopic, shardSubnetAttestationTopic} from "../util";
 import {NetworkRpc} from "./rpc";
