+++ test/e2e/sync/rpc.test.ts
@@ -2,11 +2,11 @@ import {expect} from "chai";
 import sinon from "sinon";
 import BN from "bn.js";
 
+import {Method} from "../../../src/constants";
 import {SyncRpc} from "../../../src/sync/rpc";
 import {ReputationStore} from "../../../src/sync/reputation";
 import {Libp2pNetwork, INetworkOptions} from "../../../src/network";
 import {BeaconDB, LevelDbController} from "../../../src/db";
-import {Method} from "../../../src/network/codec";
 
 import {MockBeaconChain} from "../../utils/mocks/chain/chain";
 import {createNode} from "../../unit/network/libp2p/util";
